Moved building run into component
