+++ src/apollo.js
@@ -7,33 +7,13 @@ const cache = new InMemoryCache();
 
 Vue.use(VueApollo)
 
-const flowcells = () => {
-  let flowcells = []
-
-  for (let position of [1,2,3,4,5]) {
-    flowcells.push({
-      __typename: "Flowcell",
-      position: position,
-      libraryName: '',
-    })
-  }
-  return flowcells
-}
-
-cache.writeData({
-  data: {
-    run: {
-      __typename: 'Run',
-      id: 'new',
-      flowcells: flowcells(),
-    },
-  },
-})
-
 const resolvers = {
   Mutation: {
     updateFlowcell: (_, { position, libraryName }) => {
       return { position, libraryName }
+    },
+    buildCovidRun: (_, { runId, flowcells }) => {
+      return { runId, flowcells }
     }
   }
 }
