+++ src/graphql/client/queries/BuildCovidRun.mutation.gql
@@ -1,0 +1,3 @@
+mutation ($runId: Int, $flowcells: [Flowcell]) {
+  buildCovidRun(runId: $runId, flowcells: $flowcells) @client
+}
